src/txts_unit/stamper/requirements.txt
importlib-metadata==4.11.3
Twisted==22.10.0
python-dotenv==0.21.0